Adds ModulesAnalyzer struct and module, cleanups a little bit
