+++ src/lib.rs
@@ -1,6 +1,8 @@
 use std::error::Error;
 use std::path::Path;
 
+pub mod file;
+pub mod modules_analyzer;
 pub mod project_scanner;
 pub mod use_cases;
